+++ packages/vector_graphics_codec/CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 1.1.6
+
+- Add support for encoding control points as IEEE 754-2008 half precision
+  floating point values.
+- Increase minimum SDK to 2.17.0.
+
 ## 1.1.4
 
 - Support more image formats and malformed MIME types.